requirements.txt
streamlit
opencv-python
easyocr
sympy
pillow
numpy
python-dotenv
google-generativeai
reportlab
python-docx
matplotlib